Added BaseModuleControlFactory to provide access for module factory specific code
