+++ DNN Platform/DotNetNuke.Web.Mvc/MvcModuleControlFactory.cs
@@ -18,23 +18,39 @@
 
 using System.Web.UI;
 using DotNetNuke.Entities.Modules;
+using DotNetNuke.Services.Localization;
 using DotNetNuke.UI.Modules;
 
 namespace DotNetNuke.Web.Mvc
 {
-    public class MvcModuleControlFactory : IModuleControlFactory
+    public class MvcModuleControlFactory : BaseModuleControlFactory
     {
-        public Control CreateControl(TemplateControl containerControl, string controlKey, string controlSrc)
+        public override Control CreateControl(TemplateControl containerControl, string controlKey, string controlSrc)
         {
             return new MvcHostControl(controlKey);
         }
 
-        public Control CreateModuleControl(TemplateControl containerControl, ModuleInfo moduleConfiguration)
+        public override Control CreateModuleControl(TemplateControl containerControl, ModuleInfo moduleConfiguration)
         {
             return new MvcHostControl();
         }
 
-        public Control CreateSettingsControl(TemplateControl containerControl, ModuleInfo moduleConfiguration, string controlSrc)
+        public override ModuleControlBase CreateModuleControl(ModuleInfo moduleConfiguration)
+        {
+            ModuleControlBase moduleControl = base.CreateModuleControl(moduleConfiguration);
+
+            var segments = moduleConfiguration.ModuleControl.ControlSrc.Replace(".mvc", "").Split('/');
+
+            moduleControl.LocalResourceFile = string.Format("~/DesktopModules/MVC/{0}/{1}/{2}.resx",
+                               moduleConfiguration.DesktopModule.FolderName,
+                               Localization.LocalResourceDirectory,
+                               segments[0]);
+
+            return moduleControl;
+
+        }
+
+        public override Control CreateSettingsControl(TemplateControl containerControl, ModuleInfo moduleConfiguration, string controlSrc)
         {
             return new MvcSettingsControl();
         }
